refactor: breaking up some large files into smaller ones
