+++ .gitignore
@@ -7,3 +7,4 @@ node_modules/
 dev.db
 **/vite.config.ts.timestamp-*.mjs
 .DS_Store
+.cursor
