+++ packages/document-drive/src/read-mode/index.ts
@@ -1,5 +1,6 @@
 import { Document } from "document-model/document";
-import { DocumentDriveServerConstructor, RemoteDriveOptions } from "../server";
+import { RemoteDriveOptions } from "../server";
+import { DocumentDriveServerConstructor } from "../server/base-server";
 import { logger } from "../utils/logger";
 import { ReadDriveSlugNotFoundError } from "./errors";
 import { ReadModeService } from "./service";
