+++ packages/document-drive/src/server/listener/listener-manager.ts
@@ -19,31 +19,7 @@ import {
   SynchronizationUnitQuery,
 } from "../types";
 import { StrandUpdateSource } from "./transmitter/types";
-
-function debounce<T extends unknown[], R>(
-  func: (...args: T) => Promise<R>,
-  delay = 250,
-) {
-  let timer: number;
-  return (immediate: boolean, ...args: T) => {
-    if (timer) {
-      clearTimeout(timer);
-    }
-    return new Promise<R>((resolve, reject) => {
-      if (immediate) {
-        func(...args)
-          .then(resolve)
-          .catch(reject);
-      } else {
-        timer = setTimeout(() => {
-          func(...args)
-            .then(resolve)
-            .catch(reject);
-        }, delay) as unknown as number;
-      }
-    });
-  };
-}
+import { debounce } from "./util";
 
 export class ListenerManager implements IListenerManager {
   static LISTENER_UPDATE_DELAY = 250;
